# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Conference Registration API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api"
    }
  ],
...
  "paths": {
...
    "/auth/user": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "",
        "description": "",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "",
        "description": "",
        "operationId": "updateUser",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/lectures/{id}/to-favorites": {
      "post": {
        "tags": [
          "Lectures"
        ],
        "summary": "",
        "description": "",
        "operationId": "toFavorites",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/lectures/{id}/remove-from-favorites": {
      "post": {
        "tags": [
          "Lectures"
        ],
        "summary": "",
        "description": "",
        "operationId": "removeFromFavorites",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Ukraine (Україна)"
          },
          "iso2": {
            "type": "string",
            "example": "UA"
          },
          "dial_code": {
            "type": "string",
            "example": "380"
          },
          "phone_mask": {
            "type": "string",
            "example": "+380 ## ### ####"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "firstname": {
            "type": "string",
            "example": "John"
          },
          "lastname": {
            "type": "string",
            "example": "James"
          },
          "type": {
            "type": "string",
            "enum": [
              "admin",
              "listener",
              "announcer"
            ]
          },
          "birthdate": {
            "type": "string",
            "format": "date"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "phone": {
            "type": "string",
            "example": "+1 (555) 555-5555"
          },
          "email": {
            "type": "string",
            "example": "user+1@groupbwt.com"
          },
          "email_verified_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "conference_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          }
        }
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "example": "John"
          },
          "lastname": {
            "type": "string",
            "example": "James"
          },
          "birthdate": {
            "type": "string",
            "format": "date"
          },
          "country_id": {
            "type": "integer",
            "example": 1
          },
          "phone": {
            "type": "string",
            "example": "+1 (555) 555-5555"
          },
          "email": {
            "type": "string",
            "example": "user+1@groupbwt.com"
          },
          "old_password": {
            "type": "string",
            "example": "12345678",
            "minLength": 8
          },
          "new_password": {
            "type": "string",
            "example": "qwerty123456",
            "minLength": 8
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}